requirements.txt
numpy
pandas
scipy
matplotlib
tqdm
cupy-cuda12x  # IMPORTANT: Users must match this to their CUDA version (11x or 12x)